ProbeDesign/openDesign/requirement.txt
pysam
biopython
tqdm
numpy
pandas
pyfaidx
nupack

blastn

